+++ core-network/src/main/java/com/materip/core_network/networkDI/NetworkModule.kt
@@ -1,9 +1,8 @@
 package com.materip.core_network
 
-import android.view.PixelCopy.request
 import com.google.gson.GsonBuilder
 import com.materip.core_database.TokenManager
-import com.materip.core_network.service.test.LoginService
+import com.materip.core_network.service.login.LoginService
 import com.skydoves.sandwich.retrofit.adapters.ApiResponseCallAdapterFactory
 import dagger.Module
 import dagger.Provides
@@ -65,7 +64,7 @@ object NetworkModule {
 
     @Singleton
     @Provides
-    fun provideLoginService(retrofit: Retrofit): LoginService{
+    fun provideLoginService(retrofit: Retrofit): LoginService {
         return retrofit.create(LoginService::class.java)
     }
 }
